steering-wheel-app/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pyjwt>=2.9.0
sqlalchemy>=2.0.0
bcrypt>=4.0.0
psycopg[binary]>=3.2.0
pymysql>=1.1.0
